implemented a test method to enter winners and updates their funds after potential payouts

+++ database/schema.sql
@@ -25,7 +25,7 @@ CREATE TABLE "bets" (
   "winner" boolean,
   "betType" text,
   "betAmount" int NOT NULL,
-  "payout" int,
+  "payout" NUMERIC,
   "placedAt" TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
